WIP: Add support for handling Issues as well.
Some refactoring so PullRequest* inherits from
Issue*.

Add a --dry-run mode for testing and read-only operation.

Add a TriageIssue and TriagePullRequest classes, that implement
some of the issue type specific triage workflow.

Add a Issue and PullRequest class for the Triage classes to use.

Add support for looking for tracebacks in Issues via the 'tbgrep'
module.