misc refactoring
Mostly related to the confusing name 'issue'.
Issues are an Issue, but don't have an issue.
Pull requests are sort of an Issue, but do have an issue.

Some of the previous self.instance that was used here, has been
replaced with a self.issue property.

This also adds a ansible version parser.